+++ lib/str.h
@@ -51,6 +51,6 @@ extern STATUS _str_copy(str *t, str *s);
 
 extern str _str_long_to_str(long num);
 
-extern str _str_int_to_str(long num);
+extern str _str_int_to_str(int num);
 
 #endif //GAMES_SCORE_STATISTICS_STR_H
